refactor: move getTrayPosition + TrayPosition enum into tray-app
getTrayPosition() and the TrayPosition enum now live in composenativetray-app
(package ...trayapp), next to the popup positioning they feed, so the core
artifact exposes no tray-corner / positioning API at all. The native
tray-region queries are reached through typed wrappers
(WindowsTrayInitializer.notificationIconsRegion, MacTrayInitializer.statusItemRegion)
so the JNI bridges stay internal to core.

Core's TrayPosition.kt is reduced to the raw click tracker (TrayClickTracker +
TrayClickPoint) and renamed TrayClickTracker.kt. Demos and README updated for
the new package. This also removes the Linux core-only degradation: getTrayPosition
now always has Tao geometry available in tray-app.
